+++ vee/port/core/inc/LLDEVICE_linux_configuration.h
@@ -1,0 +1,32 @@
+/*
+ * C
+ *
+ * Copyright 2024 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ */
+
+#ifndef  LLDEVICE_LINUX_CONFIGURATION_H
+#define  LLDEVICE_LINUX_CONFIGURATION_H
+
+/**
+ * @file
+ * @brief LLDEVICE API port for Linux configuration.
+ * @author MicroEJ Developer Team
+ */
+
+#ifdef __cplusplus
+	extern "C" {
+#endif
+
+
+/**
+ * @brief The platform architecture name.
+ */
+#define LLDEVICE_ARCHITECTURE	"Linux"
+
+
+#ifdef __cplusplus
+	}
+#endif
+
+#endif // LLDEVICE_LINUX_CONFIGURATION_H